fixat med admin men inte optimalt än
