+++ src/ShowInDom.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { Outlet } from 'react-router-dom'
 import MenuNavbar from './components/MenuNavbar'
-import Admin from './pages/admin'
 import Footer from './components/Footer'
