+++ src/components/scss/MenuNavbar.scss
@@ -75,7 +75,7 @@ nav{
     background-color: black;
     min-width: 10rem;
     align-items: center;
-    right: 1rem;
+    right: 2rem;
 
     a{
         padding:0.5rem;
