+++ src/pages/admin/scss/addproduct.css
