+++ src/pages/admin/scss/admin.scss
@@ -1,0 +1,6 @@
+
+
+.renderAdminPages{
+    width: 100%;
+}
+
